fn phráe 2"

+++ admin/api/status.php
@@ -1,35 +1,41 @@
 <?php
-	include "config.php";
+include "config.php";
 
-	$result = 0;
-	$table = (!empty($_POST['table'])) ? htmlspecialchars($_POST['table']) : '';
-	$id = (!empty($_POST['id'])) ? htmlspecialchars($_POST['id']) : 0;
-	$attr = (!empty($_POST['attr'])) ? htmlspecialchars($_POST['attr']) : '';
+use Tuezy\Config;
+use Tuezy\SecurityHelper;
 
-	if($id)
-	{
-		$status_detail = $d->rawQueryOne("select status from #_$table where id = $id limit 0,1");
-		$status_array = (!empty($status_detail['status'])) ? explode(',', $status_detail['status']) : array();
+// Initialize Config
+$configObj = new Config($config);
 
-		if(array_search($attr, $status_array) !== false)
-		{
-			$key = array_search($attr, $status_array);
-			unset($status_array[$key]);
-		}
-		else
-		{
-			array_push($status_array, $attr);
-		}
+$result = 0;
+$table = SecurityHelper::sanitizePost('table', '');
+$id = (int)SecurityHelper::sanitizePost('id', 0);
+$attr = SecurityHelper::sanitizePost('attr', '');
 
-		$data = array();
-		$data['status'] = (!empty($status_array)) ? implode(',', $status_array) : "";
-		$d->where('id', $id);
-		if($d->update($table, $data))
-		{
-			$result = 1;
-			$cache->delete();
-		}
+if ($id && $table && $attr) {
+	$status_detail = $d->rawQueryOne(
+		"SELECT status FROM #_{$table} WHERE id = ? LIMIT 0,1",
+		[$id]
+	);
+	
+	$status_array = (!empty($status_detail['status'])) ? explode(',', $status_detail['status']) : [];
+
+	if (($key = array_search($attr, $status_array)) !== false) {
+		unset($status_array[$key]);
+	} else {
+		$status_array[] = $attr;
 	}
 
-	echo $result;
+	$data = [
+		'status' => (!empty($status_array)) ? implode(',', $status_array) : ""
+	];
+	
+	$d->where('id', $id);
+	if ($d->update($table, $data)) {
+		$result = 1;
+		$cache->delete();
+	}
+}
+
+echo $result;
 ?>
